Renamed Aux.js to ReactAux.js
