+++ src/components/signUpSheet/SignUpSheetComponent.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import { connect } from 'react-redux'
 import * as actions from '../../redux'
 import { NavLink } from 'react-router-dom';
-import Aux from '../../hoc/ReactAux/Aux'
+import Aux from '../../hoc/ReactAux/ReactAux'
 import { Loading } from '../UI/spinner/LoadingComponent';
 import { Alert } from 'reactstrap';
 class SignUpSheetComponent extends Component {
