+++ src/components/submittedContentEdit/SubmittedContentEditComponent.js
@@ -15,7 +15,7 @@ import React, {Component} from 'react';
 import { connect } from 'react-redux'
 import * as actions from '../../redux'
 import { NavLink } from 'react-router-dom';
-import Aux from '../../hoc/ReactAux/Aux'
+import Aux from '../../hoc/ReactAux/ReactAux'
 import { Loading } from '../UI/spinner/LoadingComponent';
 import { Alert } from 'reactstrap';
 class SubmittedContentEditComponent extends Component {
