+++ src/hoc/ReactAux/ReactAux.js
